investwell-backend-master/investwell/requirements.txt
asgiref==3.8.1
backports.functools-lru-cache==1.6.1
boto3==1.16.47
botocore==1.19.63
Brotli==1.0.9
CacheControl==0.12.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
charset-normalizer==2.1.1
click==8.0.0
ConfigArgParse==1.2.3
cryptography==3.4.8
datadog==0.39.0
decorator==4.4.2
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.3
djangorestframework==3.15.2
EasyProcess==0.3
filelock==3.16.1
Flask==2.1.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==23.9.1
geventhttpclient==1.5.1
gunicorn==23.0.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.1
jmespath==0.10.0
locust==2.7.0
MarkupSafe==2.0.1
msgpack==1.0.0
packaging==21.3
platformdirs==4.3.6
psutil==5.6.7
pycparser==2.20
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyVirtualDisplay==0.2.4
ratelimitingfilter==1.2
requests==2.28.1
roundrobin==0.0.4
s3transfer==0.3.7
selenium==3.141.0
setuptools-scm==7.0.2
six==1.15.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.25.10
virtualenv==20.27.1
Werkzeug==2.0.3
whitenoise==6.8.2
zope.event==4.5.0
zope.interface==5.1.2
